+++ src/assets/SettingsIcon.tsx
@@ -15,7 +15,6 @@ const SettingsIcon = () => (
       <g mask="url(#mask0_2675_1559)">
         <path
           d="M14.2594 3.24755C15.0882 3.46144 15.8711 3.7898 16.59 4.21495C16.59 4.21495 16.9363 3.8685 17.3253 3.47949C17.6078 3.19674 17.9911 3.03816 18.3907 3.03816C18.7897 3.03816 19.1731 3.19674 19.4556 3.47949C19.7982 3.82219 20.1778 4.20174 20.5205 4.54444C20.8033 4.82686 20.9618 5.21021 20.9618 5.60935C20.9618 6.00891 20.8033 6.3922 20.5205 6.67463C20.1315 7.06364 19.785 7.41005 19.785 7.41005C20.2102 8.12892 20.5385 8.91178 20.7525 9.74063H21.7906C22.6224 9.74063 23.2969 10.4151 23.2969 11.2469V12.7531C23.2969 13.5849 22.6224 14.2594 21.7906 14.2594H20.7525C20.5385 15.0882 20.2102 15.8711 19.785 16.5899C19.785 16.5899 20.1315 16.9364 20.5205 17.3253C20.8033 17.6078 20.9618 17.9911 20.9618 18.3907C20.9618 18.7898 20.8033 19.1731 20.5205 19.4556C20.1778 19.7982 19.7982 20.1778 19.4556 20.5205C19.1731 20.8033 18.7897 20.9618 18.3907 20.9618C17.9911 20.9618 17.6078 20.8033 17.3253 20.5205C16.9363 20.1315 16.59 19.785 16.59 19.785C15.8711 20.2102 15.0882 20.5386 14.2594 20.7524V21.7906C14.2594 22.6224 13.5849 23.2969 12.7531 23.2969H11.2469C10.415 23.2969 9.74063 22.6224 9.74063 21.7906V20.7524C8.91183 20.5386 8.12892 20.2102 7.41005 19.785C7.41005 19.785 7.06364 20.1315 6.67463 20.5205C6.3922 20.8033 6.00886 20.9618 5.60934 20.9618C5.2102 20.9618 4.82681 20.8033 4.54444 20.5205C4.20178 20.1778 3.82219 19.7982 3.47949 19.4556C3.19674 19.1731 3.03816 18.7898 3.03816 18.3907C3.03816 17.9911 3.19674 17.6078 3.47949 17.3253C3.8685 16.9364 4.21495 16.5899 4.21495 16.5899C3.7898 15.8711 3.46148 15.0882 3.24755 14.2594H2.20941C1.37756 14.2594 0.703125 13.5849 0.703125 12.7531V11.2469C0.703125 10.4151 1.37756 9.74063 2.20941 9.74063H3.24755C3.46148 8.91178 3.7898 8.12892 4.21495 7.41005C4.21495 7.41005 3.8685 7.06364 3.47949 6.67463C3.19674 6.3922 3.03816 6.00891 3.03816 5.60935C3.03816 5.21021 3.19674 4.82686 3.47949 4.54444C3.82219 4.20174 4.20178 3.82219 4.54444 3.47949C4.82681 3.19674 5.2102 3.03816 5.60934 3.03816C6.00886 3.03816 6.3922 3.19674 6.67463 3.47949C7.06364 3.8685 7.41005 4.21495 7.41005 4.21495C8.12892 3.7898 8.91183 3.46144 9.74063 3.24755V2.20936C9.74063 1.37752 10.415 0.703127 11.2469 0.703127H12.7531C13.5849 0.703127 14.2594 1.37752 14.2594 2.20936V3.24755Z"
-          stroke="#58667E"
           strokeWidth="1.2"
           strokeMiterlimit="10"
           strokeLinecap="round"
@@ -23,7 +22,6 @@ const SettingsIcon = () => (
         />
         <path
           d="M12.0002 7.4813C14.4942 7.4813 16.5189 9.50606 16.5189 12C16.5189 14.494 14.4942 16.5188 12.0002 16.5188C9.50621 16.5188 7.48145 14.494 7.48145 12C7.48145 9.50606 9.50621 7.4813 12.0002 7.4813Z"
-          stroke="#58667E"
           strokeWidth="1.2"
           strokeMiterlimit="10"
           strokeLinecap="round"
